+++ custom_model_runner/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### [1.16.6] - 2025-02-26
+##### Changed
+- vLLM and NIM predictors now allow the user to specify `model` name as part of a chat completion request. The old behavior hard-coded `datarobot-deployed-llm` as the model name. While this value is still the default if no model name is passed in the completion request, it can be overriden by setting the `served_model_name` runtime parameter.
+- Adds support for `score()` hook in `NIMPredictor()` class.
+
 #### [1.16.5] - 2025-02-18
 ##### Changed
 - vLLM predictor now supports `default_parallelism` runtime parameter for setting either -tp or -pp
@@ -12,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Extend DRUM API to support proxy requests to the model server.
 - Extend NimPredictor to support multi-container deployments
 - Internal adjustments to support newer versions of Keras.
-- Adds support for `predict_unstructured()` hook in `NIMPredictor()` class.
+- Adds support for `score_unstructured()` hook in `NIMPredictor()` class.
 
 #### [1.16.4] - 2025-01-31
 ##### Changed